+++ CHANGELOG.md
@@ -6,10 +6,13 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [1.0.2] - 2023-09-28
+- Add double check before apply and delete manifest file.
+
+## [1.0.1] - 2023-08-10
+- Change default keybinding.
+
 ## [1.0.0] - 2023-08-03
 - Use Webpack to bundle src code.
 - Update devDependencies (typescript 3.8.3 --> 5.1.3; vscode 1.43.0 --> 1.80.0)
 - Add Funtion: Sync Online resource config to Local YAML.
-
-## [1.0.1] - 2023-08-10
-- Change default keybinding.